Universally sovereign for all
You were all free to travel wherever you want as long as you are following the universal law and you are not hurting anybody you are not causing harm you and source are in alignment. Nobody can get can give you a ticket. Nobody can do commerce with you. You are a sovereign soul.